feat, refactor: create StationManagerView module
